* Moved Argument Parsing to main * Renamed cgjProfileTool according to Swift Name Schemes * Removed SwiftPM dependency from CgjProfileCore => Make CgjProfileCore usable on it's own
